requirements.txt
pandas
numpy
scikit-learn
xgboost
joblib
flask
supabase
python-dotenv
python-docx
imbalanced-learn
gunicorn
